// src/app/models/BonoCarbonoInfo.ts
export interface BonoCarbonoInfo {
    id: number;
    owner: string;
    generador: string;
    nombreGenerador: string;
    timestampGenerado: number;
    toneladasCO2: number;
    fuenteEnergia: string;
    vendido: boolean;
    comprador: string;
    nombreComprador: string;
    timestampVenta: number;
  }